added tjs.textEncode() and tjs.textDecode() as alternative UTF-8 API
The original TextEncoder/TextDecoder pollyfill used in txikijs
(https://www.npmjs.com/package/text-encoding) was large and impacted startup
times significantly.

It was replaced with a UTF-8 only version
(https://www.npmjs.com/package/text-encoding-utf-8), but that had performance
issues due to inefficient array manipulation not present in the original.
Some improvements from the original can be patched in, but it is still
a slow option.

As an alternative, tjs.textEncode() and tjs.textDecode() expose the native
QuickJS encoding functions which performs order of magnitudes faster, e.g.
in a test loading and decoding a 112kb file:

    68.17 ms for TextEncoder().decode() from "improved" text-encoding-utf-8
     0.62 ms for tjs.textDecode() native implementation